Move 'Employee' (and related) Models to 'employee' app
Move the Model classes to the employee app and prepare for new migrations
